Rename json_dict module to types
