+++ tamr_client/attributes/attribute.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, field, replace
 from typing import Optional
 
 import tamr_client as tc
-from tamr_client.json_dict import JsonDict
+from tamr_client.types import JsonDict
 
 _RESERVED_NAMES = frozenset(
     [
