--- tamr_client/json_dict.py
@@ -1,4 +1,0 @@
-# taken from https://github.com/python/typing/issues/182
-from typing import Any, Dict
-
-JsonDict = Dict[str, Any]
