+++ tamr_client/types.py
@@ -1,0 +1,4 @@
+from typing import Any, Dict
+
+# taken from https://github.com/python/typing/issues/182
+JsonDict = Dict[str, Any]
